Added EatItemsGoal
- Forgot to commit this file on the first commit
